+++ spec/support/test_jobs.rb
@@ -1,0 +1,15 @@
+class DefaultJob
+  @queue = :default
+
+  def self.perform
+    # do something
+  end
+end
+
+class FailJob
+  @queue = :default
+
+  def self.perform
+    raise "I'm a failure"
+  end
+end