+++ CHANGELOG.md
@@ -2,4 +2,8 @@
 
 ## [0.1.0] - 2025-02-13
 
-- Initial release
+- Add jobs_pending metric
+- Add jobs_processed metric
+- Add jobs_failed metric
+- Add workers_total metric
+- Add workers_working metric